+++ cgi-bin/sitemap.js
@@ -9,12 +9,63 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-const algoliasearch = require('algoliasearch');
+const pick = require('lodash.pick');
+const pickBy = require('lodash.pickby');
 const { logger } = require('@adobe/openwhisk-action-logger');
 const { wrap } = require('@adobe/openwhisk-action-utils');
 const Downloader = require('@adobe/helix-pipeline/src/utils/Downloader.js');
 const { MountConfig, IndexConfig } = require('@adobe/helix-shared');
 const { getOriginalHost } = require('../src/utils');
+
+const algolia = require('../src/providers/algolia.js');
+const azure = require('../src/providers/azure.js');
+
+/**
+ * List of known index providers.
+ */
+const providers = [
+  algolia, azure,
+];
+
+/**
+ * Create the search provider that will return all paths for the sitemap.
+ *
+ * @param {object} index index configuration
+ * @param {object} params parameters relevant for the provider
+ * @returns provider instance
+ */
+function createSearchProvider(index, params) {
+  const { owner, repo, downloader } = params;
+
+  const { sitemap, target } = index;
+  if (sitemap) {
+    return {
+      search: async (query, {
+        hitsPerPage,
+        page,
+        attributesToRetrieve,
+      }) => {
+        const uri = new URL(index.fetch.replace(/\{owner\}/g, owner).replace(/\{repo\}/g, repo).replace(/\{path\}/g, sitemap));
+        uri.searchParams.append('limit', hitsPerPage);
+        uri.searchParams.append('offset', page * hitsPerPage);
+        const res = await downloader.fetch({ uri: uri.toString() });
+        if (res.status !== 200) {
+          return {};
+        }
+        const hits = JSON.parse(res.body);
+        return {
+          hits: hits.map((hit) => pick(hit, attributesToRetrieve)),
+        };
+      },
+    };
+  }
+
+  const candidate = providers.find((c) => c.match(target));
+  if (!candidate) {
+    throw new Error(`No search provider match for target: ${target}`);
+  }
+  return candidate.create(index, params);
+}
 
 /**
  * Return the location element containing the absolute path to a hit in the index.
@@ -43,8 +94,6 @@ async function run(params) {
     __hlx_ref: ref,
     page = 0,
     hitsPerPage = 100,
-    ALGOLIA_API_KEY,
-    ALGOLIA_APP_ID,
     __ow_headers: headers,
     __ow_logger: log,
   } = params;
@@ -58,12 +107,6 @@ async function run(params) {
   if (!ref) {
     throw new Error('__hlx_ref parameter missing.');
   }
-  if (!ALGOLIA_API_KEY) {
-    throw new Error('ALGOLIA_API_KEY parameter missing.');
-  }
-  if (!ALGOLIA_APP_ID) {
-    throw new Error('ALGOLIA_APP_ID parameter missing.');
-  }
   const coords = { owner, repo, ref };
   const action = {
     request: { params: coords },
@@ -75,7 +118,7 @@ async function run(params) {
   };
 
   const downloader = new Downloader({}, action, {
-    forceHttp1: ALGOLIA_APP_ID === 'foo', // use http1 for tests
+    forceHttp1: owner === 'me', // use http1 for tests
   });
   let files;
 
@@ -126,10 +169,26 @@ async function run(params) {
   }
 
   // TODO: when there are multiple indices, find the one appropriate for sitemap generation
-  const algolia = algoliasearch(ALGOLIA_APP_ID, ALGOLIA_API_KEY);
-  const indexname = `${owner}--${repo}--${Object.keys(config.indices)[0]}`;
-  const index = algolia.initIndex(indexname);
-  const result = await index.search('', {
+  const firstIndexName = Object.keys(config.indices)[0];
+  let provider;
+
+  try {
+    provider = createSearchProvider(config.indices[firstIndexName], {
+      owner,
+      repo,
+      ref,
+      downloader,
+      ...pickBy(params, (value, key) => /^[A-Z_]+$/.test(key)),
+    });
+  } catch (e) {
+    log.error('Unable to create search provider', e);
+    return {
+      statusCode: 500,
+      reason: e.message,
+    };
+  }
+
+  const result = await provider.search('', {
     hitsPerPage,
     page,
     attributesToRetrieve: ['path', 'external-path'],