+++ src/html.pre.js
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 const jquery = require('jquery');
-const { getAbsoluteUrl } = require('../src/utils');
+const { getAbsoluteUrl } = require('./utils.js');
 
 /**
  * The 'pre' function that is executed before the HTML is rendered
@@ -81,9 +81,7 @@ function pre(context) {
     .toArray();
   meta.description = `${desc.slice(0, 25).join(' ')}${desc.length > 25 ? ' ...' : ''}`;
   meta.url = getAbsoluteUrl(request.headers, request.url);
-  // switched from absolute to relative URL for the canonical link, see https://github.com/adobe/helix-pages/issues/284
-  meta.canonicalUrl = request.url;
-  meta.imageUrl = getAbsoluteUrl(request.headers, content.image);
+  meta.imageUrl = getAbsoluteUrl(request.headers, content.image || '/default-meta-image.png');
 }
 
 module.exports.pre = pre;